Add files via upload
Topic: Stock price prediction using LSTM

Predicting the stock prices in the future is an important use-case in finance. If we can build a model with reasonable accuracy which can help us in predicting the movement of the stock prices for an asset, it can yield good returns.

The Long Short Term Memory (LSTM) networks are a refined version of Recurrent Neural Network (RNN) which perform well on such problems.

Let us use the LSTM network to predict stock prices for a security.